xwayland: Add support for EGLDevices
To quote the spec page for EGL_EXT_device_base:

	The EGL device family of extensions offers a standardized
	framework for bootstrapping EGL without the use of any underlying
	"native" APIs or functionality.

EGL_EXT_device_base is the part of the API which defines both the
extensions for querying and enumerating available EGL devices. Both
querying and enumerating are handled by the EGL library, and are not as
platform specific as other methods such as using wl_drm to query
devices.

We add support for this in order to allow easy querying of both GBM
based and nvidia rendering devices. We still leave defaulting to GBM
through wl_drm provided devices as a fallback for systems without
support for these extensions.

Signed-off-by: Lyude Paul <lyude@redhat.com>